refactor: 주문 상세 엔티티 패키지 변경
